refactor: rename package 'subcmd' to 'gateway'
