+++ gateway/builder/base.go
@@ -12,10 +12,10 @@ import (
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/vim-volt/volt/fileutil"
+	"github.com/vim-volt/volt/gateway/buildinfo"
 	"github.com/vim-volt/volt/lockjson"
 	"github.com/vim-volt/volt/logger"
 	"github.com/vim-volt/volt/pathutil"
-	"github.com/vim-volt/volt/subcmd/buildinfo"
 )
 
 // BaseBuilder is a base struct which all builders must implement
