+++ gateway/builder/builder.go
@@ -5,9 +5,9 @@ import (
 	"os"
 
 	"github.com/vim-volt/volt/config"
+	"github.com/vim-volt/volt/gateway/buildinfo"
 	"github.com/vim-volt/volt/logger"
 	"github.com/vim-volt/volt/pathutil"
-	"github.com/vim-volt/volt/subcmd/buildinfo"
 )
 
 // Builder creates/updates ~/.vim/pack/volt directory
