+++ gateway/disable.go
@@ -1,4 +1,4 @@
-package subcmd
+package gateway
 
 import (
 	"flag"
@@ -10,29 +10,29 @@ import (
 )
 
 func init() {
-	cmdMap["enable"] = &enableCmd{}
+	cmdMap["disable"] = &disableCmd{}
 }
 
-type enableCmd struct {
+type disableCmd struct {
 	helped bool
 }
 
-func (cmd *enableCmd) ProhibitRootExecution(args []string) bool { return true }
+func (cmd *disableCmd) ProhibitRootExecution(args []string) bool { return true }
 
-func (cmd *enableCmd) FlagSet() *flag.FlagSet {
+func (cmd *disableCmd) FlagSet() *flag.FlagSet {
 	fs := flag.NewFlagSet(os.Args[0], flag.ContinueOnError)
 	fs.SetOutput(os.Stdout)
 	fs.Usage = func() {
 		fmt.Print(`
 Usage
-  volt enable [-help] {repository} [{repository2} ...]
+  volt disable [-help] {repository} [{repository2} ...]
 
 Quick example
-  $ volt enable tyru/caw.vim # will enable tyru/caw.vim plugin in current profile
+  $ volt disable tyru/caw.vim # will disable tyru/caw.vim plugin in current profile
 
 Description
   This is shortcut of:
-  volt profile add {current profile} {repository} [{repository2} ...]` + "\n\n")
+  volt profile rm {current profile} {repository} [{repository2} ...]` + "\n\n")
 		//fmt.Println("Options")
 		//fs.PrintDefaults()
 		fmt.Println()
@@ -41,7 +41,7 @@ Description
 	return fs
 }
 
-func (cmd *enableCmd) Run(cmdctx *CmdContext) *Error {
+func (cmd *disableCmd) Run(cmdctx *CmdContext) *Error {
 	reposPathList, err := cmd.parseArgs(cmdctx.Args)
 	if err == ErrShowedHelp {
 		return nil
@@ -51,7 +51,7 @@ func (cmd *enableCmd) Run(cmdctx *CmdContext) *Error {
 	}
 
 	profCmd := profileCmd{}
-	err = profCmd.doAdd(append(
+	err = profCmd.doRm(append(
 		[]string{"-current"},
 		reposPathList.Strings()...,
 	))
@@ -62,7 +62,7 @@ func (cmd *enableCmd) Run(cmdctx *CmdContext) *Error {
 	return nil
 }
 
-func (cmd *enableCmd) parseArgs(args []string) (pathutil.ReposPathList, error) {
+func (cmd *disableCmd) parseArgs(args []string) (pathutil.ReposPathList, error) {
 	fs := cmd.FlagSet()
 	fs.Parse(args)
 	if cmd.helped {
