+++ gateway/help_test.go
@@ -1,4 +1,4 @@
-package subcmd
+package gateway
 
 import (
 	"strings"
