+++ gateway/self_upgrade.go
@@ -1,4 +1,4 @@
-package subcmd
+package gateway
 
 import (
 	"encoding/json"
