+++ main.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/vim-volt/volt/config"
+	"github.com/vim-volt/volt/gateway"
 	"github.com/vim-volt/volt/lockjson"
 	"github.com/vim-volt/volt/logger"
-	"github.com/vim-volt/volt/subcmd"
 )
 
 func main() {
@@ -39,7 +39,7 @@ func run(args []string) (int, string) {
 		return 1, err.Error()
 	}
 
-	c := subcmd.LookUpCmd(subCmd)
+	c := gateway.LookUpCmd(subCmd)
 	if c == nil {
 		return 3, "Unknown command '" + subCmd + "'"
 	}
@@ -62,7 +62,7 @@ func run(args []string) (int, string) {
 		return 30, errors.Wrap(err, "failed to read config.toml").Error()
 	}
 
-	result := c.Run(&subcmd.CmdContext{
+	result := c.Run(&gateway.CmdContext{
 		Args:     args,
 		LockJSON: lockJSON,
 		Config:   cfg,
@@ -70,16 +70,16 @@ func run(args []string) (int, string) {
 	return result.Code, result.Msg
 }
 
-func expandAlias(subCmd string, args []string) (string, []string, error) {
+func expandAlias(gateway string, args []string) (string, []string, error) {
 	cfg, err := config.Read()
 	if err != nil {
 		return "", nil, errors.Wrap(err, "could not read config.toml")
 	}
-	if newArgs, exists := cfg.Alias[subCmd]; exists && len(newArgs) > 0 {
-		subCmd = newArgs[0]
+	if newArgs, exists := cfg.Alias[gateway]; exists && len(newArgs) > 0 {
+		gateway = newArgs[0]
 		args = append(newArgs[1:], args...)
 	}
-	return subCmd, args, nil
+	return gateway, args, nil
 }
 
 // On Windows, this function always returns nil.
